* Create HKCSensorCoordinator at platform level * Log tracebacks for coordinator update exceptions * Remove datetime.min as it is timezone naive * Remove @Throttle as it doesn't work
