+++ custom_components/hkc_alarm/__init__.py
@@ -1,10 +1,10 @@
 import logging
+import traceback
 from pyhkc.hkc_api import HKCAlarm
 from datetime import datetime, timezone, timedelta
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import callback, HomeAssistant
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
-from homeassistant.util import Throttle
 from .const import DOMAIN, DEFAULT_UPDATE_INTERVAL, CONF_UPDATE_INTERVAL
 
 _logger = logging.getLogger(__name__)
@@ -18,7 +18,7 @@ class HKCAlarmCoordinator(DataUpdateCoordinator):
             name="hkc_alarm_data",
             update_interval=timedelta(seconds=update_interval),
         )
-        self._last_update = datetime.min
+        self._last_update = None
         self._hkc_alarm = hkc_alarm
         self.panel_time_offset = None
         self.status = None
@@ -26,7 +26,6 @@ class HKCAlarmCoordinator(DataUpdateCoordinator):
         # self.sensor_data = None
 
     async def _async_update_data(self):
-        @Throttle(timedelta(seconds=30))
         def fetch_data():
             self.status = self._hkc_alarm.get_system_status()
             self.panel_data = self._hkc_alarm.get_panel()
@@ -34,12 +33,43 @@ class HKCAlarmCoordinator(DataUpdateCoordinator):
 
         try:
             now = datetime.now(timezone.utc)
-            if now > self._last_update + timedelta(seconds=30):
+            if self._last_update is None or now > self._last_update + timedelta(seconds=30):
                 self._last_update = now
                 await self.hass.async_add_executor_job(fetch_data)
             return self.status, self.panel_data #, self.sensor_data
         except Exception as e:
             _logger.error(f"Exception occurred while fetching HKC data: {e}")
+            _logger.error(traceback.format_exc())
+            raise UpdateFailed(f"Failed to update: {e}")
+
+class HKCSensorCoordinator(DataUpdateCoordinator):
+    def __init__(self, hass: HomeAssistant, config_entry: ConfigEntry, hkc_alarm: HKCAlarm, alarm_coordinator: DataUpdateCoordinator, update_interval) -> None:
+        super().__init__(
+            hass,
+            _logger,
+            config_entry=config_entry,
+            name="hkc_sensor_data",
+            update_interval=timedelta(seconds=update_interval),
+        )
+        self._last_update = None
+        self._hkc_alarm = hkc_alarm
+        self._alarm_coordinator = alarm_coordinator
+        self.sensor_data = None
+
+    async def _async_update_data(self):
+        def fetch_data():
+            self.sensor_data = self._hkc_alarm.get_all_inputs()
+
+        try:
+            await self._alarm_coordinator.async_refresh()
+            now = datetime.now(timezone.utc)
+            if self._last_update is None or now > self._last_update + timedelta(seconds=30):
+                self._last_update = now
+                await self.hass.async_add_executor_job(fetch_data)
+            return self.sensor_data
+        except Exception as e:
+            _logger.error(f"Exception occurred while fetching HKC data: {e}")
+            _logger.error(traceback.format_exc())
             raise UpdateFailed(f"Failed to update: {e}")
 
 
@@ -54,19 +84,25 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
 
     update_interval = entry.options.get(CONF_UPDATE_INTERVAL, DEFAULT_UPDATE_INTERVAL)
     alarm_coordinator = HKCAlarmCoordinator(hass, entry, hkc_alarm, update_interval)
+    sensor_coordinator = HKCSensorCoordinator(
+        hass, entry, hkc_alarm, alarm_coordinator, update_interval
+    )
     await alarm_coordinator.async_config_entry_first_refresh()
+    await sensor_coordinator.async_config_entry_first_refresh()
 
     hass.data.setdefault(DOMAIN, {})[entry.entry_id] = {
         "hkc_alarm": hkc_alarm,
         "update_interval": update_interval,
         "alarm_coordinator": alarm_coordinator,
+        "sensor_coordinator": sensor_coordinator,
     }
 
     @callback
     def update_options(updated_entry: ConfigEntry) -> None:
         update_interval = updated_entry.options.get(CONF_UPDATE_INTERVAL, DEFAULT_UPDATE_INTERVAL)
         hass.data[DOMAIN][updated_entry.entry_id]["update_interval"] = update_interval
-        ## TODO: update active coordinator intervals
+        alarm_coordinator.update_interval = timedelta(seconds=update_interval)
+        sensor_coordinator.update_interval = timedelta(seconds=update_interval)
 
     entry.add_update_listener(update_options)
